fix long frames in ZMTP 1.0
Summary:
If the size is 255 that means that the real size is in the next 8 bytes
as a long. Problem is previously didn't check if the buffer has 8 more
bytes to read. This returns -1 which signals that the length isn't yet
readable and will cause parseZMTP1Header to return false. This will
cause parseMessage to resetReaderIndex back to the mark and to wait for
more bytes.